Heb de fouten er uit gehaald en 'm verplaatst naar een andere map, lijkt mij logischer
