+++ Main-project/DataModels/Flight.cs
@@ -11,14 +11,16 @@ namespace Main_project.DataModels
         public string DepartureCity { get; set; }
         public string DestinationCity { get; set; }
 
-        public Flight(string number, string departure, string destination, DateTime departuretime, int seats, decimal price)
+        public Flight(string number, string departure, string destination, DateTime departuretime, int seats, decimal price, string departureCity, string destinationCity)
         {
             FlightNumber = number;
-            Departure = departure;
-            Destination = destination;
+            DepartureAirportCode = departure;
+            DestinationAirportCode = destination;
             DepartureTime = departuretime;
             SeatsAvailable = seats;
             Price = price;
+            DepartureCity = departureCity;
+            DestinationCity = destinationCity;
         }
 
         // Formatting class data
